removed incorrect refresh logic and testing variables
Previous refresh logic does not work in all browsers (notably Chrome, as
Chrome sets the HTTP_CACHE_CONTROL in the $_SERVER settings on the first
load, as opposed to Firefox and IE which set those on refresh); as I
found no all-encompassing method for dealing with refresh, I've
determined I'll just let the browser post the page back with the same
variable (which will do the opposite action that had just been
performed).  So if something is added to cart and page is refreshed,
that same item will be removed from cart.  I may revisit the refresh
problem in the future, but right now I'd prefer the page to work in all
browsers.